# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plato's Core — Signal Command Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Workspace {
  id: string;
  name: string;
  plan: string;
  quota: number;
  used: number;
}

export interface SourceHealth {
  name: string;
  yield: number;
  status: string;
}

export interface Dashboard {
  rawScanned: number;
  uniqueAccounts: number;
  duplicatesSuppressed: number;
  staleSignals: number;
  billableOpportunities: number;
  intentUpdates: number;
  quota: number;
  used: number;
  sources: SourceHealth[];
}

export interface Contact {
  name: string;
  title: string;
  linkedin: string;
}

export interface Signal {
  id: string;
  company: string;
  contact: Contact;
  source: string;
  sourcePlatform: string;
  sourceUrl: string;
  evidenceSnippet: string;
  whyNow: string;
  fitScore: number;
  confidenceScore: number;
  freshnessScore: number;
  seenAt: string;
  lastVerifiedAt: string;
  disposition: string;
  recommendedChannel: string;
  owner: string;
  route: string;
  crmStatus: string;
  dedupeStatus: string;
  modelPath: string;
  rawSource: string;
  /** @nullable */
  feedback?: string | null;
}

export interface AssignInput {
  owner: string;
}

export interface FeedbackInput {
  signalId: string;
  workspaceId: string;
  feedback: string;
}