# lib/api-zod/src/generated/types/assignInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plato's Core — Signal Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignInput {
  owner: string;
}