# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plato's Core — Signal Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  name: string;
  title: string;
  linkedin: string;
}