# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plato's Core — Signal Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceHealth } from './sourceHealth';

export interface Dashboard {
  rawScanned: number;
  uniqueAccounts: number;
  duplicatesSuppressed: number;
  staleSignals: number;
  billableOpportunities: number;
  intentUpdates: number;
  quota: number;
  used: number;
  sources: SourceHealth[];
}